requirements.txt
pymarkovchain
discord.py
dotenv